requirements.txt
Django~=4.0.2
django-allauth~=0.48.0
pillow~=9.0.1